CET-4384 post pr fixes
